[Ref #713] Extract SearchListenerHelper to its own file from MainViews
